% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution.R
\name{find_vol_i}
\alias{find_vol_i}
\title{Find the volume of a stock solution to aliquot}
\usage{
find_vol_i(conc_i, conc_f, vol_f)
}
\arguments{
\item{conc_i}{Initial or stock solution concentration.}

\item{conc_f}{New (or final) solution concentration.}

\item{vol_f}{New (or final) solution volume.}
}
\value{
A numeric vector whose values represent the volume of stock solution
to use. The units of these volumes are to be interpreted as the same as in
the final volume \code{vol_f}.
}
\description{
Suppose you need to prepare a new working solution from a stock solution
whose concentration is \code{conc_i}. If the final concentration and volume are
\code{conc_f} and \code{vol_f}, then \code{\link[=find_vol_i]{find_vol_i()}} determines the volume of stock
solution to use (\code{vol_i}), i.e. the aliquot volume. \code{\link[=find_vol_i]{find_vol_i()}} uses the
well-known \href{https://en.wikipedia.org/wiki/Dilution_(equation)}{Dilution Equation}:

\deqn{c_i v_i = c_f v_f}

Thus, the initial or stock volume to aliquot is:

\deqn{v_i = \frac{c_f v_f}{c_i}}
}
\section{Admonition}{
This function has no means of checking the correct usage of units in
concentrations and volumes. The user needs to ensure that the units are
coherent, i.e. use the same unit in \code{conc_i} and \code{conc_f}, e.g. M or uM; and
expect the same units as \code{vol_f} in the final result.
}

\examples{
# Stock solution concentration of 10 Molar (`conc_i`).
# Desired/new solution concentration of 5 Molar (`conc_f`).
# Final volume of the new solution: 1 Litre (`vol_f`).
find_vol_i(conc_i = 10, conc_f = 5, vol_f = 1)

# `find_vol_i()` is vectorised and applies tidyverse recycling
find_vol_i(conc_i = 10, conc_f = c(5, 5/2, 5/4, 5/8), vol_f = 1)

}
