% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protbot-package.R
\docType{package}
\name{protbot-package}
\alias{protbot}
\alias{protbot-package}
\title{protbot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://life-science-protocols.github.io/protbot/}
  \item \url{https://github.com/life-science-protocols/protbot/}
  \item Report bugs at \url{https://github.com/life-science-protocols/protbot/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Mendes-Silva \email{lmesilva@pattern.institute} (\href{https://orcid.org/0000-0002-5989-637X}{ORCID})

Authors:
\itemize{
  \item Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
