% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_calibration.R
\name{read_sample_calibration}
\alias{read_sample_calibration}
\title{Read a sample calibration file}
\usage{
read_sample_calibration(file = stop("Please provide a `file` path"))
}
\arguments{
\item{file}{A string with the sample calibration file name.}
}
\value{
A data frame resulting from the reading of the sample calibration
file pointed by \code{file}.
}
\description{
This function reads a calibration file that is bundled with protbot.
}
